Implemented ResponsesIterator for CommandDispatcher
Another in a series of commits to simplify mpv manager and it's command disptacher,
this one introduces synchronous ResponsesIterator - it's responsible for
reading responses on mpv socket and present them to a client in an interator
fashion. This is done, in order to get rid of a separate loop in a goroutine,
which was hard to manage and hard to get errors from.