Directory probing refactor
Removed unnecessary go routines, simplifying the code in question.
As a result calling code (REST handlers etc.) should ensure that adding and
probing of the directories and files is done in necessary fashion.
(REST handlers are already run in a different go routine, so no additional
creation of go routines is necessary - only main calling "AddDirectories"
needs to ensure that it's done concurrently)