Saving of playlist current entry idx on playlist change (#18)
Due to implementation of a subscriber pattern in mpv states, it's now possible for server to observe on the state changes just as a sse server would. This enables watching on various playlist-related changes on a state to be separated from particular state-related event handlers (I don't want playlist-related logic to run in a handler of mpv property change, because changes to the playlist will also happen from the REST/other APIs requests). As such a stub of functionality is implemented with this commit - changes to the current named (read from file) playlist are saved on a change of playlists playback. The next step would be to also save currently played playlist state on common signals.